backend/requirements.txt
# SEPE Templates Comparator - Production Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# HTTP Client & Web Scraping
requests==2.31.0
httpx==0.25.2

# Configuration Management
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Task Queue (for background processing)
celery[redis]==5.3.4

# Email Support
# smtplib is part of Python standard library

# Utilities
python-dateutil==2.8.2
pillow==10.1.0
